+++ src/workspaceviewlabel.c
@@ -191,8 +191,9 @@ workspaceviewlabel_name_edit_activate(iEntry *self, gpointer user_data)
 
 	g_autofree char *text = NULL;
 	g_object_get(self, "text", &text, NULL);
-	if (text)
-		workspace_rename(ws, text);
+	if (text &&
+		!workspace_rename(ws, text))
+		workspace_show_error(ws);
 
 	g_object_set(wviewlabel, "edit", FALSE, NULL);
 }